Add keyboard hook to wait for [Numpad ENTER]
The keyboard hook merely listens to a keydown of the numpad enter.
If that occurs, you may or may not specify the tool to invoke the normal
[ENTER] as well.

Due to limitation in .net, I can not hook and throw away the [Numpad ENTER].
It is only possible to hook KEYBOARD_LL and MOUSE_LL globally. Other hooks
do not work.

This makes the filtering annoying and turns this tool basically in a keylogger.
So AV might call us out. :(

Alternative for this: Make a native library and call the hooks from there
(or do everything in another language..)