Rework the process grabbing mechanism
Okay, here's what happened:
The old method of grabbing a process that is creating while RatherWeird is running was implemented by listening to ObjectCreate messages. This had the advantage that the associated process was done loading. The drawback is, that this is triggered so often, that it shreds Computers.

The solution is to (boringly) check processes each second and report the result to the caller.
Then the caller checks with `IsWindow` if the process has a window.
If that is the case, we continue and set the process.

Additionally: I rewrote the memory part of the application.
I still have an open feature in another branch so I'll keep both Memory implementations for now. But the new one automatically caches the `OpenProcess` requests so that the caller doesnt need to give a damn.